ARACHNE-2719  Update AD/LDAP user import routines to work with Authenticator. fix PR remarks - change signature of the authenticate method - extract BaseAuthenticationServiceImpl class
